data/web/requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-keycloak-auth==0.9.5
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
mariadb==1.0.11
Markdown==3.3.6
mysql==0.0.3
mysqlclient==2.1.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
django-cors-headers==3.11.0
